Applied patch [ 1876926 ] avoid truncation of auto scale by Wim Lewis, should fix issues with 0 sized output of autooptimizer

+++ src/tools/autooptimiser.cpp
@@ -224,11 +224,11 @@ int main(int argc, char *argv[])
         opts.setHFOV(fitPano.getResultHorizontalFOV());
         opts.setHeight(roundi(fitPano.getResultHeight()));
         vfov = opts.getVFOV();
-		hfov = opts.getHFOV();
+        hfov = opts.getHFOV();
         // avoid perspective projection if field of view > 100 deg
         double mf = 100;
         if (vfov < mf) {
-        // cylindrical or rectilinear
+            // cylindrical or rectilinear
             if (hfov < mf) {
                 opts.setProjection(PanoramaOptions::RECTILINEAR);
             } else {
@@ -240,8 +240,8 @@ int main(int argc, char *argv[])
         double sizeFactor = 0.7;
 
         pano.setOptions(opts);
-		int w = CalculateOptimalScale::calcOptimalScale(pano);
-		opts.setWidth(roundi(opts.getWidth()*w*sizeFactor), true);
+        double w = CalculateOptimalScale::calcOptimalScale(pano);
+        opts.setWidth(roundi(opts.getWidth()*w*sizeFactor), true);
         pano.setOptions(opts);
     }
